tiktok-platform-microservices/services/device-service/requirements.txt
# Device Service Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.3

# Redis Streams
redis==5.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
PyJWT==2.8.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For TestClient
